+++ insanic/authentication/handlers.py
@@ -61,12 +61,13 @@ def jwt_service_decode_handler(token):
     )
 
 
-def jwt_service_payload_handler(service):
+def jwt_service_payload_handler(service, user):
     payload = {
         "source": settings.SERVICE_NAME,
         "aud": service.service_name,
         "source_ip": socket.gethostbyname(socket.gethostname()),
-        "destination_version": "0.0.1"
+        "destination_version": "0.0.1",
+        "user": user
     }
     return payload