+++ insanic/conf/global_settings.py
@@ -41,6 +41,8 @@ CACHES = {
     },
 }
 
+TASK_CONTEXT_REQUEST_USER = "request_user"
+
 REDIS_HOST = "localhost"
 REDIS_PORT = 6379
 REDIS_DB = 0